requirements.txt
# Python version 2.7.11 is used in this project

pytest==2.9.2               # Library to write test cases for this projects
python-dateutil==2.4.2      # Library to validate datetime strings
